Create symlinks for files differently than for folders on Windows
Using junctions for symlinks to files does not work. The symlink is
created by trying to read the file using the type command results in
"Access is denied."

Try different options on for the mklink command:
* no option or /D results in "You do not have sufficient privilege to
perform this operation."
* existing /J option does not work
* option /H works for files.

Update win32_symlink/2 to check file type of Source and use options /H
and /J for files and folders respectively.