Skip write_ini when {include_erts, false}
When `{include_erts, false}` the `erts_dir` will most likely by
found in the `C:\Program Files` folder. In that case Windows will
complain "Access is denied" when writing to the `erl.exe` file.

Writting (updating) the ini file is useful when `{include_erts, true}`
so the release folder can be relocated to any location with ease.

So, only skip the `:write_ini` step when `erts_dir` is not found in
the `release_root_dir`, but found in `:set_erts_dir_from_erl`.