Get Windows release to work with {include_erts, false}
A release built with {include_erts, false} does not run since it cannot find
locally defined apps.

The local apps are in the boot script in the form:
  "$ROOT/lib/myapp-0.1.0/ebin"
this form works for the builtin apps (kernel, stdlib, ...) but not the local
apps.

Add a boot script variable RELEASE_DIR only for {include_erts, false} on
Windows. The boot script form changes to:
  "$RELEASE_DIR/lib/myapp-0.1.0/ebin"

Windows script templates are updated to provide the variable value
  -boot_var RELEASE_DIR "%release_root_dir%"